Update instance.hpp
- Make the int go long long for count_request

+++ src/instance.hpp
@@ -61,7 +61,7 @@ struct InstanceData {
     Request* requests; // Array containing all the requests
     Cache* caches; // Array containing all the caches
     bool** cache_affectation; // Matrix of boolean to describe the affectation of video to cache. Shape: VxC (row: video, col: cache)
-    int sum_request_count; // Sum of all count in subsequent requests
+    long long int sum_request_count; // Sum of all count in subsequent requests
     long long score; // Weighted score: sum over requests of gain * count
 };
